move structure type
